Move interop declarations to own files. No functional changes
