+++ GitUI/CommandsDialogs/FormFileHistoryController.cs
@@ -1,3 +1,4 @@
+using System;
 using System.IO;
 using System.Text;
 using GitCommands.Utils;
