+++ GitUI/Interops/COLORREF.cs
@@ -1,0 +1,12 @@
+﻿namespace System
+{
+    internal static partial class NativeMethods
+    {
+        public struct COLORREF
+        {
+            public byte R;
+            public byte G;
+            public byte B;
+        }
+    }
+}
