+++ GitUI/Interops/DTT_CALLBACK_PROC.cs
@@ -1,0 +1,11 @@
+using System.Runtime.InteropServices;
+
+namespace System
+{
+    internal static partial class NativeMethods
+    {
+        public delegate int DTT_CALLBACK_PROC(IntPtr hdc,
+            [MarshalAs(UnmanagedType.LPWStr)] string text, int textLen, ref RECT rc, int flags,
+            IntPtr lParam);
+    }
+}
