+++ GitUI/Interops/Kernel32/GetShortPathName.cs
@@ -1,0 +1,11 @@
+using System.Runtime.InteropServices;
+using System.Text;
+
+namespace System
+{
+    internal static partial class NativeMethods
+    {
+        [DllImport(Libraries.Kernel32, ExactSpelling = true, CharSet = CharSet.Auto, SetLastError = true)]
+        public static extern uint GetShortPathName(string lpszLongPath, StringBuilder lpszShortPath, int cchBuffer);
+    }
+}
