+++ GitUI/Interops/Libraries.cs
@@ -1,0 +1,11 @@
+﻿namespace System
+{
+    internal static partial class Libraries
+    {
+        public const string Gdi32 = "gdi32.dll";
+        public const string Kernel32 = "kernel32.dll";
+        public const string User32 = "user32.dll";
+        public const string UxTheme = "uxtheme.dll";
+        public const string WinInet = "wininet.dll";
+    }
+}
