+++ GitUI/Interops/TEXTSHADOWTYPE.cs
@@ -1,0 +1,12 @@
+﻿namespace System
+{
+    internal static partial class NativeMethods
+    {
+        public enum TEXTSHADOWTYPE : int
+        {
+            None = 0,
+            Single = 1,
+            Continuous = 2,
+        }
+    }
+}
