+++ GitUI/Interops/User32/GetActiveWindow.cs
@@ -1,0 +1,10 @@
+using System.Runtime.InteropServices;
+
+namespace System
+{
+    internal static partial class NativeMethods
+    {
+        [DllImport(Libraries.User32)]
+        public static extern IntPtr GetActiveWindow();
+    }
+}
