+++ GitUI/Interops/UxTheme/CloseThemeData.cs
@@ -1,0 +1,10 @@
+using System.Runtime.InteropServices;
+
+namespace System
+{
+    internal static partial class NativeMethods
+    {
+        [DllImport(Libraries.UxTheme, ExactSpelling = true)]
+        public static extern int CloseThemeData(IntPtr hTheme);
+    }
+}
