+++ GitUI/Interops/UxTheme/GetThemeColor.cs
@@ -1,0 +1,10 @@
+using System.Runtime.InteropServices;
+
+namespace System
+{
+    internal static partial class NativeMethods
+    {
+        [DllImport(Libraries.UxTheme, ExactSpelling = true)]
+        public static extern int GetThemeColor(IntPtr hTheme, int iPartId, int iStateId, int iPropId, out COLORREF pColor);
+    }
+}
