+++ GitUI/Interops/UxTheme/OpenThemeData.cs
@@ -1,0 +1,10 @@
+using System.Runtime.InteropServices;
+
+namespace System
+{
+    internal static partial class NativeMethods
+    {
+        [DllImport(Libraries.UxTheme, ExactSpelling = true, CharSet = CharSet.Unicode)]
+        public static extern IntPtr OpenThemeData(IntPtr hWnd, string classList);
+    }
+}
