+++ GitUI/Interops/UxTheme/SetWindowTheme.cs
@@ -1,0 +1,10 @@
+using System.Runtime.InteropServices;
+
+namespace System
+{
+    internal static partial class NativeMethods
+    {
+        [DllImport(Libraries.UxTheme, CharSet = CharSet.Unicode)]
+        public static extern int SetWindowTheme(IntPtr hWnd, string pszSubAppName, string pszSubIdList);
+    }
+}
