+++ GitUI/Theming/ThemeManager.cs
@@ -255,7 +255,7 @@ namespace GitUI.Theming
 
             foreach (Form form in Application.OpenForms)
             {
-                NativeMethods.SendMessageInt(form.Handle, NativeConstants.WM_SYSCOLORCHANGE,
+                NativeMethods.SendMessageInt(form.Handle, NativeMethods.WM_SYSCOLORCHANGE,
                     IntPtr.Zero, IntPtr.Zero);
             }
         }
