+++ GitUI/UserControls/ExListView.cs
@@ -9,6 +9,7 @@ using System.Runtime.InteropServices;
 using System.Windows.Forms;
 using GitCommands.Utils;
 using JetBrains.Annotations;
+using static System.NativeMethods;
 
 namespace GitUI.UserControls
 {
@@ -66,12 +67,12 @@ namespace GitUI.UserControls
             };
 
             var handleRef = new HandleRef(this, Handle);
-            if (NativeMethods.SendMessage(handleRef, NativeMethods.LVM_SUBITEMHITTEST, (IntPtr)(-1), ref info) == new IntPtr(-1))
+            if (SendMessage(handleRef, LVM_SUBITEMHITTEST, (IntPtr)(-1), ref info) == new IntPtr(-1))
             {
                 return null;
             }
 
-            if ((info.flags & NativeMethods.LVHITTESTFLAGS.LVHT_EX_GROUP_HEADER) == 0)
+            if ((info.flags & LVHITTESTFLAGS.LVHT_EX_GROUP_HEADER) == 0)
             {
                 return null;
             }
@@ -81,7 +82,7 @@ namespace GitUI.UserControls
                 var groupId = GetGroupId(group);
                 if (info.iItem == groupId)
                 {
-                    bool isCollapseButton = (info.flags & NativeMethods.LVHITTESTFLAGS.LVHT_EX_GROUP_COLLAPSE) > 0;
+                    bool isCollapseButton = (info.flags & LVHITTESTFLAGS.LVHT_EX_GROUP_COLLAPSE) > 0;
                     return new ListViewGroupHitInfo(group, isCollapseButton, location);
                 }
             }
@@ -136,19 +137,19 @@ namespace GitUI.UserControls
             var message = m;
             switch (m.Msg)
             {
-                case NativeMethods.WM_LBUTTONUP when GetGroupHitInfo(message.LParam.ToPoint())?.IsCollapseButton == true:
+                case WM_LBUTTONUP when GetGroupHitInfo(message.LParam.ToPoint())?.IsCollapseButton == true:
                     DefWndProc(ref m); // collapse / expand by clicking button in group header
                     break;
 
-                case NativeMethods.LVM_INSERTGROUP:
+                case LVM_INSERTGROUP:
                     base.WndProc(ref m);
                     HandleAddedGroup(m);
                     break;
 
-                case NativeMethods.WM_RBUTTONUP when IsGroupMouseEventHandled(MouseButtons.Right, isDown: false):
-                case NativeMethods.WM_RBUTTONDOWN when IsGroupMouseEventHandled(MouseButtons.Right, isDown: true):
-                case NativeMethods.WM_LBUTTONUP when IsGroupMouseEventHandled(MouseButtons.Left, isDown: false):
-                case NativeMethods.WM_LBUTTONDOWN when IsGroupMouseEventHandled(MouseButtons.Left, isDown: true):
+                case WM_RBUTTONUP when IsGroupMouseEventHandled(MouseButtons.Right, isDown: false):
+                case WM_RBUTTONDOWN when IsGroupMouseEventHandled(MouseButtons.Right, isDown: true):
+                case WM_LBUTTONUP when IsGroupMouseEventHandled(MouseButtons.Left, isDown: false):
+                case WM_LBUTTONDOWN when IsGroupMouseEventHandled(MouseButtons.Left, isDown: true):
                     break;
 
                 default:
@@ -164,18 +165,18 @@ namespace GitUI.UserControls
 
                 switch (msg.Msg)
                 {
-                    case NativeMethods.WM_VSCROLL:
+                    case WM_VSCROLL:
                         type = (ScrollEventType)LowWord(msg.WParam.ToInt64());
                         newValue = HighWord(msg.WParam.ToInt64());
                         break;
 
-                    case NativeMethods.WM_MOUSEWHEEL:
+                    case WM_MOUSEWHEEL:
                         type = HighWord(msg.WParam.ToInt64()) > 0
                             ? ScrollEventType.SmallDecrement
                             : ScrollEventType.SmallIncrement;
                         break;
 
-                    case NativeMethods.WM_KEYDOWN:
+                    case WM_KEYDOWN:
                         switch ((Keys)msg.WParam.ToInt32())
                         {
                             case Keys.Up:
@@ -206,7 +207,7 @@ namespace GitUI.UserControls
                         return;
                 }
 
-                newValue = newValue ?? NativeMethods.GetScrollPos(Handle, NativeMethods.SB_VERT);
+                newValue = newValue ?? GetScrollPos(Handle, SB_VERT);
                 Scroll?.Invoke(this, new ScrollEventArgs(type, newValue.Value));
 
                 short LowWord(long number) =>
@@ -257,12 +258,12 @@ namespace GitUI.UserControls
                     var lvgroup = new NativeMethods.LVGROUP();
                     lvgroup.CbSize = Marshal.SizeOf(lvgroup);
                     lvgroup.State = state;
-                    lvgroup.Mask = NativeMethods.ListViewGroupMask.State;
+                    lvgroup.Mask = ListViewGroupMask.State;
                     lvgroup.IGroupId = groupId;
 
                     var handleRef = new HandleRef(this, Handle);
 
-                    NativeMethods.SendMessage(handleRef, NativeMethods.LVM_SETGROUPINFO, (IntPtr)groupId, ref lvgroup);
+                    SendMessage(handleRef, LVM_SETGROUPINFO, (IntPtr)groupId, ref lvgroup);
                 }
             }
 
@@ -355,144 +356,6 @@ namespace GitUI.UserControls
             }
 
             return -1;
-        }
-
-        private static class NativeMethods
-        {
-            [DllImport("user32", CharSet = CharSet.Auto)]
-            public static extern IntPtr SendMessage(HandleRef hWnd,
-                                                   int msg,
-                                                   IntPtr wParam,
-                                                   ref LVHITTESTINFO lParam);
-
-            [DllImport("user32", CharSet = CharSet.Auto)]
-            public static extern IntPtr SendMessage(HandleRef hWnd,
-                                                   int msg,
-                                                   IntPtr wParam,
-                                                   ref LVGROUP lParam);
-
-            [DllImport("user32.dll", CharSet = CharSet.Auto)]
-            public static extern int GetScrollPos(IntPtr hWnd, int nBar);
-
-            #region Windows constants
-
-            public const int WM_LBUTTONDOWN = 0x0201;
-            public const int WM_LBUTTONUP = 0x0202;
-            public const int WM_RBUTTONDOWN = 0x0204;
-            public const int WM_RBUTTONUP = 0x0205;
-
-            public const int WM_VSCROLL = 0x115;
-            public const int WM_MOUSEWHEEL = 0x020A;
-            public const int WM_KEYDOWN = 0x0100;
-            public const int SB_VERT = 1;
-
-            public const int LVM_FIRST = 0x1000;
-            public const int LVM_HITTEST = LVM_FIRST + 18;
-            public const int LVM_SETGROUPINFO = LVM_FIRST + 147;
-            public const int LVM_SUBITEMHITTEST = LVM_FIRST + 57;
-            public const int LVM_INSERTGROUP = LVM_FIRST + 145;
-
-            #endregion
-
-            [StructLayout(LayoutKind.Sequential)]
-            public readonly struct POINT
-            {
-                public readonly int X;
-                public readonly int Y;
-
-                public POINT(int x, int y)
-                {
-                    X = x;
-                    Y = y;
-                }
-
-                public static implicit operator System.Drawing.Point(POINT p) => new System.Drawing.Point(p.X, p.Y);
-                public static implicit operator POINT(System.Drawing.Point p) => new POINT(p.X, p.Y);
-            }
-
-            [StructLayout(LayoutKind.Sequential)]
-            public struct NMHDR
-            {
-                public IntPtr hwndFrom;
-                public IntPtr idFrom;
-                public int code;
-            }
-
-            /// <summary>
-            /// see http://msdn.microsoft.com/en-us/library/bb774754%28v=VS.85%29.aspx
-            /// </summary>
-            [StructLayout(LayoutKind.Sequential)]
-            public struct LVHITTESTINFO
-            {
-                public POINT pt;
-                public LVHITTESTFLAGS flags;
-                public int iItem;
-                public int iSubItem;
-
-                // Vista/Win7+
-                public int iGroup;
-            }
-
-            /// <summary>
-            /// see http://msdn.microsoft.com/en-us/library/bb774754%28v=VS.85%29.aspx
-            /// </summary>
-            [Flags]
-            public enum LVHITTESTFLAGS : uint
-            {
-                LVHT_NOWHERE = 0x00000001,
-                LVHT_ONITEMICON = 0x00000002,
-                LVHT_ONITEMLABEL = 0x00000004,
-                LVHT_ONITEMSTATEICON = 0x00000008,
-                LVHT_ONITEM = LVHT_ONITEMICON | LVHT_ONITEMLABEL | LVHT_ONITEMSTATEICON,
-                LVHT_ABOVE = 0x00000008,
-                LVHT_BELOW = 0x00000010,
-                LVHT_TORIGHT = 0x00000020,
-                LVHT_TOLEFT = 0x00000040,
-
-                // Vista/Win7+ only
-                LVHT_EX_GROUP_HEADER = 0x10000000,
-                LVHT_EX_GROUP_FOOTER = 0x20000000,
-                LVHT_EX_GROUP_COLLAPSE = 0x40000000,
-                LVHT_EX_GROUP_BACKGROUND = 0x80000000,
-                LVHT_EX_GROUP_STATEICON = 0x01000000,
-                LVHT_EX_GROUP_SUBSETLINK = 0x02000000,
-            }
-
-            public enum ListViewGroupMask : uint
-            {
-                None = 0x00000,
-                Header = 0x00001,
-                Footer = 0x00002,
-                State = 0x00004,
-                Align = 0x00008,
-                GroupId = 0x00010,
-                SubTitle = 0x00100,
-                Task = 0x00200,
-                DescriptionTop = 0x00400,
-                DescriptionBottom = 0x00800,
-                TitleImage = 0x01000,
-                ExtendedImage = 0x02000,
-                Items = 0x04000,
-                Subset = 0x08000,
-                SubsetItems = 0x10000
-            }
-
-            [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
-            public struct LVGROUP
-            {
-                public int CbSize;
-                public ListViewGroupMask Mask;
-                [MarshalAs(UnmanagedType.LPWStr)]
-                public string PszHeader;
-                public int CchHeader;
-                [MarshalAs(UnmanagedType.LPWStr)]
-                public string PszFooter;
-                public int CchFooter;
-                public int IGroupId;
-                public int StateMask;
-                public ListViewGroupState State;
-                public uint UAlign;
-            }
         }
     }
 }
